# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * My Secrets - Premium Encrypted Vault API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  vaultLocked?: boolean;
  autoLockMinutes?: number;
  twoFactorEnabled?: boolean;
  createdAt: string;
}

export interface UserProfileUpdate {
  displayName?: string;
  autoLockMinutes?: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  color: string;
  isSystem: boolean;
  secretCount: number;
}

export interface CategoryInput {
  name: string;
  icon: string;
  color: string;
}

export interface CategoryUpdate {
  name?: string;
  icon?: string;
  color?: string;
}

export interface SecretMeta {
  id: number;
  title: string;
  categorySlug: string;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  /** @nullable */
  description?: string | null;
  tags?: string[];
  isFavorite: boolean;
  isArchived: boolean;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  lastAccessedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Secret {
  id: number;
  title: string;
  categorySlug: string;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  encryptedValue: string;
  /** @nullable */
  description?: string | null;
  tags?: string[];
  isFavorite: boolean;
  isArchived: boolean;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  lastAccessedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SecretInput {
  title: string;
  categoryId: number;
  encryptedValue: string;
  description?: string;
  tags?: string[];
  expiresAt?: string;
}

export interface SecretUpdate {
  title?: string;
  categoryId?: number;
  encryptedValue?: string;
  description?: string;
  tags?: string[];
  /** @nullable */
  expiresAt?: string | null;
}

export interface ActivityEntry {
  id: number;
  action: string;
  resourceType: string;
  /** @nullable */
  resourceId?: number | null;
  /** @nullable */
  resourceTitle?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  createdAt: string;
}

export interface CategoryCount {
  categorySlug: string;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  count: number;
}

export interface DashboardStats {
  totalSecrets: number;
  favoriteSecrets: number;
  archivedSecrets: number;
  expiringSecrets: number;
  categoryCounts: CategoryCount[];
  recentActivity: ActivityEntry[];
}

export interface SecurityCheck {
  name: string;
  passed: boolean;
  weight: number;
  description: string;
}

export interface SecurityScore {
  score: number;
  grade: string;
  checks: SecurityCheck[];
}

export type ListSecretsParams = {
category?: string;
search?: string;
tag?: string;
archived?: boolean;
favorite?: boolean;
};

export type ListActivityParams = {
limit?: number;
offset?: number;
};