# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * My Secrets - Premium Encrypted Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEntry {
  id: number;
  action: string;
  resourceType: string;
  /** @nullable */
  resourceId?: number | null;
  /** @nullable */
  resourceTitle?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  createdAt: string;
}