# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * My Secrets - Premium Encrypted Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryCount {
  categorySlug: string;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  count: number;
}