# lib/api-zod/src/generated/types/secretUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * My Secrets - Premium Encrypted Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface SecretUpdate {
  title?: string;
  categoryId?: number;
  encryptedValue?: string;
  description?: string;
  tags?: string[];
  /** @nullable */
  expiresAt?: string | null;
}