# lib/api-zod/src/generated/types/securityCheck.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * My Secrets - Premium Encrypted Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface SecurityCheck {
  name: string;
  passed: boolean;
  weight: number;
  description: string;
}