# lib/api-zod/src/generated/types/securityScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * My Secrets - Premium Encrypted Vault API
 * OpenAPI spec version: 0.1.0
 */
import type { SecurityCheck } from './securityCheck';

export interface SecurityScore {
  score: number;
  grade: string;
  checks: SecurityCheck[];
}